Finally fixed all the bugs
Fixed audio, leaderboards, car-data. Now the foundation of the game is done. I can finally just start focusing on tweaking the handling to make the game feel responsive, as well as fine tuning the drifting, difficulty, and start adding more levels and hats. I have a lot of ideas for things I want to implement, but had to put on hold while I was figuring out all the bugs and fixing random stuff.