Many changes to storage
Went back to PlayerPrefs for storing data during a game session. The problem with PlayerPrefs are that they are periodically deleted from phone, as well as reset after each patch to the game. Found a solution to store all playerprefs to a txt file and store that backend on the LootLocker server. Each login this txt file is fetched from the server and updates the playerprefs. Highscores and unlockables are already stored backend.